*: move redact to its own package and redact Range
