+++ pkg/redact/redact.go
@@ -1,6 +1,6 @@
 // Copyright 2020 PingCAP, Inc. Licensed under Apache-2.0.
 
-package logutil
+package redact
 
 import (
 	"encoding/hex"
@@ -25,13 +25,6 @@ func RedactString(arg string) string {
 		return "?"
 	}
 	return arg
-}
-
-type stringer struct{}
-
-// String implement fmt.Stringer
-func (s stringer) String() string {
-	return "?"
 }
 
 // RedactKey receives a key return omitted information if redact log enabled
