+++ pkg/rtree/rtree.go
@@ -11,6 +11,7 @@ import (
 	"github.com/pingcap/log"
 
 	"github.com/pingcap/br/pkg/logutil"
+	"github.com/pingcap/br/pkg/redact"
 )
 
 // Range represents a backup response.
@@ -22,7 +23,7 @@ type Range struct {
 
 // String formats a range to a string.
 func (rg *Range) String() string {
-	return fmt.Sprintf("[%x %x]", rg.StartKey, rg.EndKey)
+	return fmt.Sprintf("[%x %x)", redact.RedactKey(rg.StartKey), redact.RedactKey(rg.EndKey))
 }
 
 // BytesAndKeys returns total bytes and keys in a range.
